normed grid
