+++ lib/iron_goat/lib/iron_goat/lib/src/priv/grid/grid.h
@@ -1,0 +1,16 @@
+/*
+** EPITECH PROJECT, 2021
+** iron$
+** File description:
+** goat
+*/
+
+#ifndef __IRON_GOAT_GRID__H__
+    #define  __IRON_GOAT_GRID__H__
+
+    struct checker_tokens_orientation {
+        ig_grid_orientation_t orientation;
+        char *match;
+    };
+
+#endif
